added different checked strategies to the tree-select
